Combining Makefile and moving unit test to src/unit directory.
* Also removed doxygen tags.
