+++ src/Makefile
@@ -432,6 +432,9 @@ ifeq ($(USE_FAST_FLOAT),yes)
 	FINAL_CFLAGS += -D USE_FAST_FLOAT
 	ENGINE_SERVER_OBJ += fast_float/fast_float_strtod.o
 	ENGINE_CLI_OBJ += fast_float/fast_float_strtod.o
+	ENGINE_TEST_OBJ += fast_float/fast_float_strtod.o
+	CXX ?= c++
+	CXXFLAGS += -Wall -W $(OPT) $(DEBUG)
 	# Clang uses libc++ for standard c++ library vs libstdc++ used by gcc.
 	ifeq (clang,$(CLANG))
 		FINAL_LIBS += -lc++
@@ -445,8 +448,8 @@ all: $(SERVER_NAME) $(ENGINE_SENTINEL_NAME) $(ENGINE_CLI_NAME) $(ENGINE_BENCHMAR
 	@echo "Hint: It's a good idea to run 'make test' ;)"
 	@echo ""
 
-fast_float/fast_float_strtod.o:
-	cd fast_float && $(MAKE) CFLAGS=" $(CFLAGS) " LDFLAGS=" $(LDFLAGS) "
+fast_float/fast_float_strtod.o: fast_float/fast_float_strtod.cpp ../deps/fast_float/fast_float.h
+	$(QUIET_CXX)$(CXX) $(CXXFLAGS) $(CFLAGS) $(SERVER_CFLAGS) $(LDFLAGS) -D FASTFLOAT_ALLOWS_LEADING_PLUS -c $< -o $@
 
 Makefile.dep:
 	-$(SERVER_CC) -MM $(ALL_SOURCES) > Makefile.dep 2> /dev/null || true
@@ -565,7 +568,6 @@ commands.c: $(COMMANDS_DEF_FILENAME).def
 clean:
 	rm -rf $(SERVER_NAME) $(ENGINE_SENTINEL_NAME) $(ENGINE_CLI_NAME) $(ENGINE_BENCHMARK_NAME) $(ENGINE_CHECK_RDB_NAME) $(ENGINE_CHECK_AOF_NAME) $(ENGINE_UNIT_TESTS) $(ENGINE_LIB_NAME) unit/*.o unit/*.d *.o *.gcda *.gcno *.gcov valkey.info lcov-html Makefile.dep *.so
 	rm -f $(DEP)
-	cd fast_float && $(MAKE) clean
 
 .PHONY: clean
