--- src/fast_float/Makefile
@@ -1,49 +1,0 @@
-#
-# Copyright (c) 2024, Valkey contributors
-# All rights reserved.
-#
-# Redistribution and use in source and binary forms, with or without
-# modification, are permitted provided that the following conditions are met:
-#
-#   * Redistributions of source code must retain the above copyright notice,
-#     this list of conditions and the following disclaimer.
-#   * Redistributions in binary form must reproduce the above copyright
-#     notice, this list of conditions and the following disclaimer in the
-#     documentation and/or other materials provided with the distribution.
-#   * Neither the name of Redis nor the names of its contributors may be used
-#     to endorse or promote products derived from this software without
-#     specific prior written permission.
-#
-# THIS SOFTWARE IS PROVIDED BY THE COPYRIGHT HOLDERS AND CONTRIBUTORS "AS IS"
-# AND ANY EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT LIMITED TO, THE
-# IMPLIED WARRANTIES OF MERCHANTABILITY AND FITNESS FOR A PARTICULAR PURPOSE
-# ARE DISCLAIMED. IN NO EVENT SHALL THE COPYRIGHT OWNER OR CONTRIBUTORS BE
-# LIABLE FOR ANY DIRECT, INDIRECT, INCIDENTAL, SPECIAL, EXEMPLARY, OR
-# CONSEQUENTIAL DAMAGES (INCLUDING, BUT NOT LIMITED TO, PROCUREMENT OF
-# SUBSTITUTE GOODS OR SERVICES; LOSS OF USE, DATA, OR PROFITS; OR BUSINESS
-# INTERRUPTION) HOWEVER CAUSED AND ON ANY THEORY OF LIABILITY, WHETHER IN
-# CONTRACT, STRICT LIABILITY, OR TORT (INCLUDING NEGLIGENCE OR OTHERWISE)
-# ARISING IN ANY WAY OUT OF THE USE OF THIS SOFTWARE, EVEN IF ADVISED OF THE
-# POSSIBILITY OF SUCH DAMAGE.
-
-
-CXX?=c++
-# We want -O3 and -flto so that compiler can inline the function. 
-# Currently -flto fails 32-bit build, therefore removing it. 
-CXXFLAGS+=-std=c++11 -O3 -fPIC -D FASTFLOAT_ALLOWS_LEADING_PLUS
-
-%.o: %.cpp
-	$(CXX) $(CXXFLAGS) $(CFLAGS) $(LDFLAGS) -c $< -o $@
-
-.PHONY: all clean test
-
-all: fast_float_strtod.o
-
-clean:
-	rm -f *.o || true;
-	rm test || true;
-
-test: fast_float_strtod.o
-	$(CC) test_fast_float.c -c -o test_fast_float.o	
-	$(CXX) test_fast_float.o fast_float_strtod.o -o test $(CXXFLAGS)
-	./test
